backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pretorin-crm"
version = "1.0.0"
description = "API for managing sales contacts and government contract opportunities"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy==2.0.35",
    "python-jose[cryptography]==3.3.0",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "python-multipart==0.0.12",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "email-validator==2.3.0",
    "slowapi==0.1.9",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "httpx",
    "ruff",
]

[project.scripts]
pretorin-crm = "app.main:app"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
fail_under = 100
show_missing = true
exclude_lines = [
    "pragma: no cover",
]
